fdp1-unit-test: Fix buffer fill helper
The inner loop was not guarded by the outer loop as it should have been.
Fix the issue by adding relevant braces.

fdp1-buffer.c: In function ‘fdp1_fill_buffer’:
fdp1-buffer.c:32:2: error: this ‘for’ clause does not guard... [-Werror=misleading-indentation]
  for (i=0; i < buffer->n_planes; i++)
  ^~~
fdp1-buffer.c:34:3: note: ...this statement, but the latter is misleadingly indented as if it were guarded by the ‘for’
   for (k = 0; k < buffer->sizes[i]; k++) {
   ^~~
cc1: all warnings being treated as errors

Signed-off-by: Kieran Bingham <kieran@bingham.xyz>